requirements.txt
chardet==5.2.0
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.11.0.86
pandas==2.0.3
PyYAML==6.0.2
scikit_learn==1.3.2
seaborn==0.13.2
torch==2.4.1
torcheval==0.0.7
torchvision==0.19.1
